Flask/requirements.txt
requests
flask-smorest
python-dotenv